requirements.txt
streamlit
requests
google-generativeai
transformers
torch
pandas